+++ src/main/cpp/it/d4np/util/util.hpp
@@ -19,6 +19,7 @@
 #include <it/d4np/util/string_builder.hpp>
 #include <it/d4np/util/string_formatter.hpp>
 #include <it/d4np/util/string_splitter.hpp>
+#include <it/d4np/util/task_future.hpp>
 #include <it/d4np/util/type_traits.hpp>
 #include <it/d4np/util/unique_ref.hpp>
 #include <it/d4np/util/version.hpp>